refactor: Reading Session API 인증 및 검증 개선
- pauseSession, resumeSession, completeSession에 memberId 파라미터 추가
- Integer → Long 타입 통일 (totalMinutes, durationMinutes)
- ReadingSession.complete() 내부에서 totalMinutes 자동 계산
- Boolean.TRUE.equals()로 null-safe 처리
- SwaggerConfig에서 MediaType 문자열 리터럴로 변경
- ConstraintViolationException 핸들러 추가

+++ src/main/java/whoreads/backend/domain/readingsession/dto/ReadingSessionResponse.java
@@ -22,7 +22,7 @@ public class ReadingSessionResponse {
     public static class SessionDetail {
         private Long sessionId;
         private SessionStatus status;
-        private Integer totalMinutes;
+        private Long totalMinutes;
         private LocalDateTime createdAt;
         private LocalDateTime finishedAt;
         private List<IntervalInfo> intervals;
@@ -34,7 +34,7 @@ public class ReadingSessionResponse {
         private Long intervalId;
         private LocalDateTime startTime;
         private LocalDateTime endTime;
-        private Integer durationMinutes;
+        private Long durationMinutes;
     }
 
     @Getter
@@ -64,7 +64,7 @@ public class ReadingSessionResponse {
         private LocalTime startTime;
         @JsonFormat(pattern = "HH:mm")
         private LocalTime endTime;
-        private Integer totalMinutes;
+        private Long totalMinutes;
     }
 
     @Getter
